chore: extract voice and state helpers
