+++ services/ts/cephalon/src/agent/voiceContent.ts
@@ -1,0 +1,100 @@
+import { CollectionManager } from '../collectionManager';
+import { formatMessage, GenericEntry } from '../contextManager';
+import { generatePromptChoice, generateSpecialQuery } from '../util';
+import type { AIAgent } from './index';
+
+export async function generateVoiceContentFromSinglePrompt(this: AIAgent) {
+	let content = '';
+	let counter = 0;
+	const context = await this.context.compileContext([this.prompt], this.historyLimit, 5, 5, true);
+	const text = context.map((m) => m.content).join('\n');
+
+	while (!content && counter < 5) {
+		content = (await this.generateResponse({
+			specialQuery: `
+This is  a transcript of a conversation you and I have been having using a voice channel.
+${text}
+`,
+			context: [],
+		})) as string;
+		counter++;
+	}
+
+	return content;
+}
+
+export async function generateVoiceContentWithFormattedLatestmessage(this: AIAgent) {
+	let content = '';
+	let counter = 0;
+	const userCollection = this.contextManager.getCollection('transcripts') as CollectionManager<'text', 'createdAt'>;
+	const latestUserMessage = (await userCollection.getMostRecent(1))[0] as GenericEntry;
+	const context = (await this.context.compileContext([this.prompt], this.historyLimit)).filter(
+		(m) => m.content !== latestUserMessage?.text,
+	);
+
+	context.push({
+		role: 'user',
+		content: formatMessage(latestUserMessage),
+	});
+
+	while (!content && counter < 5) {
+		content = (await this.generateResponse({
+			context,
+		})) as string;
+		counter++;
+	}
+
+	return content;
+}
+
+export async function generateVoiceContentWithChoicePrompt(this: AIAgent) {
+	let content = '';
+	let counter = 0;
+	const context = await this.context.compileContext([this.prompt], this.historyLimit);
+	while (!content && counter < 5) {
+		content = (await this.generateResponse({
+			specialQuery: ` ${generatePromptChoice()} `,
+			context,
+		})) as string;
+		counter++;
+	}
+
+	return content;
+}
+
+export async function generateVoiceContentWithSpecialQuery(this: AIAgent) {
+	let content = '';
+	let counter = 0;
+	const userCollection = this.contextManager.getCollection('transcripts') as CollectionManager<'text', 'createdAt'>;
+	const latestUserMessage = (await userCollection.getMostRecent(1))[0] as GenericEntry;
+	const context = (await this.context.compileContext([this.prompt], this.historyLimit)).filter(
+		(m) => m.content !== latestUserMessage?.text,
+	);
+	while (!content && counter < 5) {
+		content = (await this.generateResponse({
+			specialQuery: generateSpecialQuery(latestUserMessage, generatePromptChoice()),
+			context,
+		})) as string;
+		counter++;
+	}
+
+	return content;
+}
+
+export async function generateVoiceContentWithoutSpecialQuery(this: AIAgent) {
+	let content = '';
+	let counter = 0;
+	const context = await this.context.compileContext([this.prompt], this.historyLimit);
+	while (!content && counter < 5) {
+		content = (await this.generateResponse({
+			context,
+		})) as string;
+		counter++;
+	}
+
+	return content;
+}
+
+export async function generateVoiceContent(this: AIAgent) {
+	return this.generateVoiceContentWithChoicePrompt();
+}
