+++ services/ts/cephalon/src/util.ts
@@ -37,3 +37,5 @@ ${promptChoice}
 Talk to the user about the image on the screen and engage with the conversation.
 `;
 }
+
+export const sleep = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
